Made some improvements

+++ move.sh
@@ -1,2 +1,23 @@
-cp .bash_aliases ~
-cp .vimrc ~ 
+yes=y
+
+if [ -f ~/.bash_aliases ];
+then
+	read -p "Overwrite ~/.bash_aliases?[y/n]: " choice
+
+	if [ "$choice" == "$yes" ];
+	then
+		echo -e "Overwriting ~/.bashrc"
+		cp .bash_aliases ~
+	fi
+fi
+
+if [ -f ~/.vimrc ];
+then
+	read -p "Overwrite ~/.vimrc?[y/n]: " choice
+	
+	if [ "$choice" == "$yes" ];
+	then
+		echo -e "Overwriting ~/.vimrc"
+		cp .vimrc ~ 
+	fi
+fi
